requirements.txt
fastapi
uvicorn
pyjwt
redis
sqlmodel
sqlalchemy
psycopg2-binary
passlib
minio
python-multipart
openai
requests
aiohttp
PyYAML